BEH: switch from JSON to pickle, add timestamp
Add a column _time_utc and a DataFrame index with the same values. This
is in case we need to reconstruct the index from JSON or other formats
w/o index support.

Remove enforcement of dtype=object in JSON IO. Switch to pickle as
default disk format. JSON was just a sad attempt to do type-preserving
IO. JSON is not made for this. Also it is a text format, which is
inefficient.