ENH: sanity checks to catch pgrid() usage errors
Catch stuff where we get passed an iterable that violates the assumption
that the iterable we get passed into pgrid() is a sequence of plists or
a sequence or some iterator over plists such as zip(a,b).

a=ps.plist(...)
b=ps.plist(...)

pgrid(a)            # instead of pgrid([a])
pgrid(zip(a,b))     # instead of pgrid([zip(a,b)])

which will now trigger errors down the line in merge_dicts() or
itr2params().

Without adding an elaborate type system, which we don't want (plists are
just lists of dicts) it is impossible to do better early usage error
checks here. Users should use pgrid([a,b]) if in doubt. For now, it is
not worth adding more code to make this convenience API even more safe
to use.