DOC: use row.to_dict() in _pset_hash_new example
row from df.iterrows() is a pandas.core.series.Series and has an
.items() method, so it behaves like a dict but is isn't one. We can
convert to a dict using dict(row) or row.to_dict().

Here duck-typing is NOT our friend, since joblib.hash(dict(row)) and
joblib.hash(row) are not the same, whereas pset_hash(row) and
pset_hash(dict(row)) are, but only because we build a dict from data in
row -- for certain combos of kwds, so not in all cases. Hence, don't
give people ideas and use row.to_dict() (or well dict(row)) in docs.